fix(go): overhaul type definitions from audit against Rust source
- Metadata: add Title, Authors, Keywords, CreatedAt, ModifiedAt, CreatedBy, ModifiedBy, Pages; remove Date
- Add DjotContent type hierarchy (9 types) with BlockType/InlineType constants
- Fix PptxMetadata fields (SlideCount/SlideNames), PageInfo (TableCount/Hidden)
- Add typed enum constants: LinkType, ImageType, TextDirection, StructuredDataType
- Fix attributes from map[string]string to [][2]string for tuple representation
- Fix integer signedness: int→uint64 for all usize-mapped fields

+++ packages/go/v4/images_test.go
@@ -256,7 +256,7 @@ func TestBatchImageExtractionFromMultiPageDocuments(t *testing.T) {
 
 	if len(result.Images) > 0 {
 		// Verify image indices are sequential and unique
-		imageIndices := make(map[int]bool)
+		imageIndices := make(map[uint64]bool)
 		for _, img := range result.Images {
 			if imageIndices[img.ImageIndex] {
 				t.Errorf("duplicate image index: %d", img.ImageIndex)
